requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
et-xmlfile==1.1.0
Flask==3.0.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
openai==1.16.2
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
pandas-stubs==2.2.1.240316
pluggy==1.4.0
pydantic==2.6.4
pydantic_core==2.16.3
pytest==8.1.1
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.2
types-pytz==2024.1.0.20240203
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
